Made OrderDetail Component
