+++ src/pages/index.js
@@ -6,7 +6,7 @@ import { RegistrationPage } from "./registration/registration";
 import { ResetPassPage } from "./resetPass/resetPass";
 import { IngredientDetailPage } from "./detail/detail";
 import { FeedPage } from "./feed/feed";
-import { FeedDetailPage } from "./feed-detail/feed-detail";
+import { OrderDetailPage } from "./order-detail/order-detail";
 import { NotFoundPage } from "./404/404";
 import { OrdersPage } from "./orders/orders";
 
@@ -19,7 +19,7 @@ export {
     RegistrationPage,
     ResetPassPage,
     FeedPage,
-    FeedDetailPage,
+    OrderDetailPage,
     NotFoundPage, 
     OrdersPage
 };
