+++ src/pages/order-detail/order-detail.module.css
@@ -1,0 +1,14 @@
+.page {
+    padding-top: 40px;
+    padding-bottom: 40px;
+    max-width: 1280px;
+    margin: 0 auto;
+}
+
+.main {
+    padding-top: 50px;;
+    width: 640px;
+    margin: 0 auto;
+    color: #F2F2F3;
+}
+
